# pp-syndication-api/src/main/resources/openapispec.yaml
openapi: 3.0.3
info:
  title: PagePulse Syndication API
  version: 1.0.0
servers:
  - url: http://localhost:8089
tags:
  - name: Documents
    description: Operations related to documents
  - name: Rules
    description: Operations related to document rules
paths:
  /documents:
    get:
      tags:
        - Documents
      summary: Get all documents
      operationId: getDocuments
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: A list of documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedDocumentApiDto'
  /rules:
    get:
      tags:
        - Rules
      summary: Get all rules
      operationId: getRules
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: A list of rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedRuleApiDto'
  /rules/{id}:
    get:
      tags:
        - Rules
      summary: Get rule by ID
      operationId: getRuleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Rule found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleApiDto'
        '404':
          description: Rule not found
components:
  schemas:
    DocumentApiDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        externalId:
          type: string
        title:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        documentLastCreatedAt:
          type: string
          format: date-time
        documentLastUpdatedAt:
          type: string
          format: date-time
    RuleApiDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
    PagedDocumentApiDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DocumentApiDto'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    PagedRuleApiDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RuleApiDto'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    PageInfo:
      type: object
      properties:
        page:
          type: integer
        pages:
          type: integer
        elements:
          type: integer
          format: int64
    PaginationApiDto:
      type: object
      properties:
        page:
          type: integer
          description: "The page number to retrieve (0-indexed)."
          default: 0
        size:
          type: integer
          description: "The number of items to retrieve per page."
          default: 10
        sort:
          type: array
          items:
            type: string
          description: "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported."
  parameters:
    page:
      in: query
      name: page
      schema:
        type: integer
        default: 0
      description: The page number to retrieve (0-indexed).
    size:
      in: query
      name: size
      schema:
        type: integer
        default: 10
      description: The number of items to retrieve per page.
    sort:
      in: query
      name: sort
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
      description: "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported."